+++ migrations/2022_07_24_195026_cities.php
@@ -12,7 +12,7 @@ class Cities extends Migration
             $table->id();
             $table->string('name_ar');
             $table->string('name_en');
-            $table->unsignedInteger('region_id');
+            $table->unsignedBigInteger('region_id');
             $table->foreign('region_id')->references('id')->on('regions')->cascadeOnDelete();
             $table->timestamps();
         });
